requirements.txt
pandas
matplotlib
shortuuid
numpy
boto3
python-dotenv
schedule